+++ devkit/src/manifest/index.ts
@@ -1,2 +1,4 @@
 export * from "./schema.js";
 export * from "./load.js";
+export * from "./language.js";
+export * from "./npm-packages.js";